🎉 feat(cpu_utilization): add CloudWatch metric alarm and SNS topic This commit adds a CloudWatch metric alarm and SNS topic to monitor EC2 CPU utilization. The `data.tf` file contains a Terraform data block that retrieves all running EC2 instances. The `main.tf` file contains a Terraform resource block that creates a CloudWatch metric alarm for each EC2 instance. The `output.tf` file contains an output block that returns the alert name for each CloudWatch metric alarm. The `sns.tf` file contains a Terraform resource block that creates an SNS topic and email subscription for each email target specified in the `var.email_targets` variable. The `checkov:skip=CKV_AWS_26: ADD REASON` comment is added to suppress a Checkov warning.
📝 chore(cpu_utlization): add variables for SNS topic, email targets, alarm threshold, evaluation periods, and environment
📝 chore(nlb): add data and locals files
This commit adds variables for the SNS topic name, email targets, alarm threshold, evaluation periods, and environment in the cpu_utilization module. These variables will be used to configure the CloudWatch alarm for CPU utilization. In addition, this commit adds data and locals files in the nlb module. These files will be used to store data and local values for the module.

🆕 feat(nlb): add CloudWatch metric alarm and SNS topic resources
This commit adds two new Terraform resources to the nlb module. The first resource is an AWS CloudWatch metric alarm that monitors the number of healthy nodes in a target group. The alarm is triggered if the number of healthy nodes falls below a specified threshold. The second resource is an AWS SNS topic that sends email notifications to a list of email targets. The email targets are specified as a variable and can be customized for each deployment.

📝 chore(nlb): add variables.tf, data.tf, and locals.tf files
This commit adds three new files to the nlb directory: variables.tf, data.tf, and locals.tf. The variables.tf file contains variables for the name of the sns topic, email targets, name of the nlb, name of the target group, and threshold for the alarm. The data.tf file contains data sources for the nlb and target group. The locals.tf file is currently empty.

🆕 feat(test): add AWS CloudWatch metric alarm and SNS topic resources
This commit adds two new Terraform resources to the test directory. The first resource is an AWS CloudWatch metric alarm that monitors the number of healthy nodes in a target group. The second resource is an AWS SNS topic that sends email notifications to specified email addresses. The AWS CloudWatch metric alarm resource is dynamic and creates an alarm for each target group specified in the var.target_name variable. The AWS SNS topic resource is also dynamic and creates a subscription for each email address specified in the var.email_targets variable.

🆕 chore(variables.tf): add sns-topic-name, email_targets, lb_name, target_name, alarm_threshold, and namespace variables to test/variables.tf
This commit adds new variables to the test/variables.tf file. These variables are used to configure the SNS topic, email targets, load balancer name, target group name, alarm threshold, and namespace for the alarm. These variables are used to make the code more flexible and easier to maintain.